+++ src/main/java/net/lopymine/mtd/exception/SuppressedMyTotemDollException.java
@@ -1,0 +1,9 @@
+package net.lopymine.mtd.exception;
+
+import java.io.IOException;
+
+public class SuppressedMyTotemDollException extends IOException {
+
+	public SuppressedMyTotemDollException() {
+	}
+}