+++ src/main/java/net/lopymine/mtd/mixin/AnvilScreenMixin.java
@@ -207,7 +207,7 @@ public abstract class AnvilScreenMixin extends ForgingScreen<AnvilScreenHandler>
 
 	@WrapOperation(at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;getName()Lnet/minecraft/text/Text;"), method = "onSlotUpdate")
 	private Text swapItemName(ItemStack stack, Operation<Text> original) {
-		if (!MyTotemDollClient.getConfig().isModEnabled() || !stack.isOf(Items.TOTEM_OF_UNDYING)) {
+		if (!MyTotemDollClient.canProcess(stack)) {
 			return original.call(stack);
 		}
 		Text customName = stack.getRealCustomName();