+++ src/main/java/net/lopymine/mtd/mixin/PlayerSkinTextureMixin.java
@@ -10,6 +10,7 @@ import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.*;
 
 
+import net.lopymine.mtd.exception.SuppressedMyTotemDollException;
 import net.lopymine.mtd.utils.mixin.MTDPlayerSkinTexture;
 import net.lopymine.mtd.utils.texture.*;