+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-- Fixed some bugs
+- Better tag menu tooltip resizing
+- Added support for other mod totems
+- Fixed stick
+- Fixed Dinnerbone
+- Fixed issue with doll registering (#18)
+- Other nice fixes and changes